--- api/guards.js
@@ -1,9 +1,0 @@
-export async function IsLogged(
-  chainable$,
-  payload,
-  context
-) {
-  if (!context.user) {
-    throw new Error('Unauthorized');
-  }
-}
