--- api/helpers/moment.js
@@ -1,5 +1,0 @@
-import moment from 'moment';
-
-export function fromNow() {
-  return moment('20111031', 'YYYYMMDD').fromNow();
-}
