--- api/interceptors.js
@@ -1,15 +1,0 @@
-import { tap } from 'rxjs/operators';
-
-export async function LoggerInterceptor(
-  chainable$,
-  root,
-  context,
-  payload,
-  descriptor
-) {
-  console.log('Before...');
-  const now = Date.now();
-  return chainable$.pipe(
-    tap(() => console.log(`After... ${Date.now() - now}ms`))
-  );
-}
