--- api/methods/test.rs
@@ -1,4 +1,0 @@
-#[no_mangle]
-pub fn add(a: i32, b: i32) -> i32 {
-  return a + b
-}
