+++ lib/app/app.module.js
@@ -27,6 +27,8 @@ const path_1 = require("path");
 const app_tokens_1 = require("./app.tokens");
 const transpile_and_load_1 = require("../helpers/transpile-and-load");
 const traverse_1 = require("../helpers/traverse/traverse");
+const traverse_map_1 = require("../helpers/traverse-map");
+const watch_bundles_1 = require("../helpers/watch-bundles");
 let AppModule = class AppModule {
 };
 AppModule = __decorate([
@@ -155,6 +157,7 @@ ${core_1.printSchema(mergedSchemas)}
                     if (config.$mode === 'advanced') {
                         yield advanced_schema_1.MakeAdvancedSchema(config, bootstrap);
                     }
+                    watch_bundles_1.watchBundles(traverse_map_1.traverseMap.map(f => f.path), config, bootstrap);
                     return true;
                 })
             }
