+++ lib/helpers/advanced-schema.js
@@ -157,6 +157,7 @@ function MakeAdvancedSchema(config, bootstrap) {
                     resolve = resolve[firstKey];
                 }
             }
+            resolve = isFunction_1.isFunction(resolve) ? resolve : () => resolve;
             bootstrap.Fields.query[resolver] = {
                 type: types[type],
                 method_name: resolver,
@@ -164,7 +165,7 @@ function MakeAdvancedSchema(config, bootstrap) {
                 public: true,
                 method_type: 'query',
                 target: () => { },
-                resolve: isFunction_1.isFunction(resolve) ? resolve : () => resolve
+                resolve
             };
         });
     });
