+++ lib/helpers/load-file.js
@@ -45,7 +45,6 @@ function loadFile(path) {
         const parent = path
             .substring(0, path.lastIndexOf('/'))
             .replace(process.cwd(), '');
-        loadedModule.constructor.$parent = parent;
         traverse_map_1.traverseMap.push({ parent, path });
         return loadedModule;
     });
