--- lib/helpers/load-yml.d.ts
@@ -1,1 +1,0 @@
-export declare function loadFile(path: string): Promise<any>;
