+++ lib/helpers/resolvers.map.d.ts
@@ -1,0 +1,1 @@
+export declare const resolversMap: Map<any, any>;
