+++ lib/helpers/resolvers.map.js
@@ -1,0 +1,3 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.resolversMap = new Map();
