--- lib/helpers/traverse.d.ts
@@ -1,2 +1,0 @@
-export declare function traverseAndLoadConfigs<T>(x: T): Promise<void>;
-export declare function recursiveIterator(obj: any): IterableIterator<any>;
