+++ lib/helpers/watch-bundles.d.ts
@@ -1,0 +1,3 @@
+import { Config } from '../app/app.tokens';
+import { BootstrapService } from '@gapi/core';
+export declare function watchBundles(paths: string[], config: Config, bootstrap: BootstrapService): void;
