+++ package.json
@@ -15,6 +15,11 @@
     "fast-glob": "^3.0.4",
     "js-yaml": "^3.13.1"
   },
+  "esm": {
+    "cjs": {
+      "cache" : false
+    }
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/rxdi/graphqj"
@@ -28,6 +33,7 @@
     "@rxdi/parcel-plugin-shebang": "0.7.5",
     "@types/graphql": "^14.2.3",
     "@types/js-yaml": "^3.12.1",
+    "moment": "^2.24.0",
     "typescript": "^3.5.3"
   },
   "types": "./lib/index.d.ts",
