--- resolvers.yml
@@ -1,26 +1,0 @@
-findUser: 💉./resolvers/findUser.resolver.yml
-findUser2: 💉./resolvers/findUser.resolver.yml
-findUser3: 💉./resolvers/findUser.resolver.yml
-findUser4: 💉./resolvers/findUser.resolver.yml
-findUser5: 💉./resolvers/findUser.resolver.yml
-findUser6: 💉./resolvers/findUser.resolver.yml
-findUser7: 💉./resolvers/findUser.resolver.yml
-findUser8: 💉./resolvers/findUser.resolver.yml
-
-findUser31: 💉./resolvers/findUser.resolver.yml
-findUser231: 💉./resolvers/findUser.resolver.yml
-findUser313: 💉./resolvers/findUser.resolver.yml
-findUse31r4: 💉./resolvers/findUser.resolver.yml
-findUse31r5: 💉./resolvers/findUser.resolver.yml
-findU31ser6: 💉./resolvers/findUser.resolver.yml
-findU31ser7: 💉./resolvers/findUser.resolver.yml
-findU31ser8: 💉./resolvers/findUser.resolver.yml
-
-find3131User: 💉./resolvers/findUser.resolver.yml
-find31User2: 💉./resolvers/findUser.resolver.yml
-fin31dUser3: 💉./resolvers/findUser.resolver.yml
-fin31dUser4: 💉./resolvers/findUser.resolver.yml
-fi3131ndUser5: 💉./resolvers/findUser.resolver.yml
-fi31ndUser6: 💉./resolvers/findUser.resolver.yml
-fin3311dUser7: 💉./resolvers/findUser.resolver.yml
-find31User8: 💉./resolvers/findUser.resolver.yml
