--- resolvers/findUser.resolver.yml
@@ -1,4 +1,0 @@
-type: User
-args:
-  payload: UserPayload
-resolve: 💉./proba/test2.js
