--- resolvers/proba/test2.js
@@ -1,10 +1,0 @@
-export async function findUser() {
-  console.log("22222222222222222222222222222222")
-  return {
-    name: 'dada',
-    email: 'dada',
-    phone: 13131,
-    arrayOfNumbers: [111, 222],
-    arrayOfStrings: ['dada', 'dada']
-  };
-}
