+++ src/app/app.module.ts
@@ -32,7 +32,7 @@ import {
 } from '../helpers/transpile-and-load';
 import { traverseAndLoadConfigs } from '../helpers/traverse/traverse';
 import { traverseMap } from '../helpers/traverse-map';
-import { watchBundles } from '../helpers/watch-files';
+import { watchBundles } from '../helpers/watch-bundles';
 
 @Module({
   imports: [VoyagerModule.forRoot()],
@@ -141,8 +141,7 @@ ${printSchema(mergedSchemas)}
       ) => {
         config = await config;
         await traverseAndLoadConfigs(config);
-        traverseMap
-        watchBundles(traverseMap.map(f => f.path))
+
         if (config.$externals) {
           const compiledPaths = await TranspileAndGetAll(
             config.$externals,
@@ -184,13 +183,10 @@ ${printSchema(mergedSchemas)}
         if (config.$mode === 'basic') {
           await MakeBasicSchema(config, bootstrap);
         }
-
         if (config.$mode === 'advanced') {
           await MakeAdvancedSchema(config, bootstrap);
         }
-        setInterval(() => {
-          console.log(config.$resolvers.findUser)
-        }, 3000)
+        watchBundles(traverseMap.map(f => f.path), config, bootstrap)
         return true;
       }
     }
