+++ src/helpers/watch-bundles.ts
@@ -1,17 +1,18 @@
 import { watch } from 'chokidar';
 import { loadFile } from './load-file';
 import { Stats } from 'fs';
-import { Container } from '@rxdi/core';
 import { Config } from '../app/app.tokens';
 import { traverseAndLoadConfigs } from './traverse/traverse';
+import { BootstrapService } from '@gapi/core';
 
-export function watchBundles(paths: string[]) {
+export function watchBundles(paths: string[], config: Config, bootstrap: BootstrapService) {
   const ignored = (p: string) => p.includes('node_modules');
-  watch(paths, { ignored }).on('change', async (path, stats: Stats) => {
-    const config = await Container.get(Config)
+  watch([...new Set(paths)], { ignored }).on('change', async (path, stats: Stats) => {
     const test = await loadFile(path)
+    console.log(test)
     await traverseAndLoadConfigs(config);
+    bootstrap.Fields.query['findUser'].resolve = function () {
 
-    console.log(test);
+    }
   });
 }
