--- test.html
@@ -1,3 +1,0 @@
-<bla-component></bla-component>
-{name} {email} {phone} {createdAt}
-A rich framework for building applications and services with GraphQL and Apollo inspired by Angular
