--- types.yml
@@ -1,6 +1,0 @@
-User:
-  name: 💉./gj.js
-  email: String => {⌛fromNow}
-  phone: Number
-  arrayOfNumbers: Number[]
-  arrayOfStrings: String[]
